ipn/{ipnauth,ipnlocal,ipnserver}: send the auth URL to the user who started interactive login
We add the ipnauth.Session interface, an optional interface that ipnauth.Actor implementations might
implement to represent a LocalAPI session, and update ipnserver.actor to implement it. The interface
facilitates linking a series of related interactions by a SessionID, such as when a notification needs to be
sent back to a specific session, rather than all active sessions, in response to a certain request.

We also add LocalBackend.WatchNotificationsAs and LocalBackend.StartLoginInteractiveAs methods,
which are like WatchNotifications and StartLoginInteractive but accept an additional parameter
specifying an ipnauth.Actor who initiates the operation. We store these actor identities in
watchSession.actor and LocalBackend.authActor, respectively,and implement LocalBackend.sendTo
and related helper methods to enable sending notifications to watchSessions associated with actors
(or, more broadly, identifiable recipients).

We then use the above to change who receives the BrowseToURL notifications:
 - For user-initiated, interactive logins, the notification is delivered only to the user
   who initiated the process. Ideally, we send it back to the same LocalAPI session that called
   StartLoginInteractive, but if there are no notification watchers associated with the caller's
   ipnauth.Session, we fall back to sending it to the same user.
   Currently, we only differentiate between users on Windows, as it is inherently a multi-user OS.
   On Unix-like and other platforms, we only distinguish between LocalAPI sessions.
 - In all other cases (e.g., node key expiration), we send the notification to all connected users.

Updates tailscale/corp#18342

Signed-off-by: Nick Khyl <nickk@tailscale.com>